feat(minijinja): deprecate litestar.contrib.minijinja and migrate to litestar.plugins.minijinja
Move ``MiniJinjaTemplateEngine`` and the related ``StateProtocol`` /
``MiniJinjaTemplate`` types to their new canonical location at
``litestar.plugins.minijinja`` and replace ``litestar.contrib.minijinja``
with a runtime-compatible deprecation shim.

This is the third (and largest, 201 LOC) of the contrib → plugins
template-engine batch (sister PRs: #4752 jinja, #4753 mako). The shim
follows the same module-level ``__getattr__`` pattern with
``version="3.0.0b0"`` / ``removal_in="3.0.0"``: bare ``import
litestar.contrib.minijinja`` is silent, attribute access emits a
``DeprecationWarning``, identity equality holds with the new module.

Coupled change: ``litestar/plugins/flash.py:62`` lazily imports
``_transform_state`` from this module. That import is migrated to the
new plugins location to avoid triggering our own deprecation warnings on
test runs. Because external callers may also depend on the old
``litestar.contrib.minijinja._transform_state`` import path, the shim
explicitly forwards the private helper too — so identity equality holds
for ``MiniJinjaTemplateEngine``, ``StateProtocol``, ``MiniJinjaTemplate``,
**and** ``_transform_state``.

Tests:

* ``tests/unit/test_plugins/test_minijinja.py`` — moved from
  ``test_contrib/`` (functional coverage of the engine itself).
* ``tests/unit/test_contrib/test_minijinja_deprecation.py`` — new tests
  pinning the shim contract for all four forwarded symbols
  (Engine, StateProtocol, Template, ``_transform_state``), identity
  equality, silent bare import, ``AttributeError`` on unknown names.

Docs:

* ``docs/reference/plugins/minijinja.rst`` added.
* Changelog and ``whats-new-3.rst`` updated.

Closes #4718. Part of the contrib → plugins umbrella (#4301).
