+++ docs/examples/plugins/flash_messages/minijinja.py
@@ -1,5 +1,5 @@
 from litestar import Litestar
-from litestar.contrib.minijinja import MiniJinjaTemplateEngine
+from litestar.plugins.minijinja import MiniJinjaTemplateEngine
 from litestar.middleware.session.server_side import ServerSideSessionConfig
 from litestar.plugins.flash import FlashConfig, FlashPlugin
 from litestar.template.config import TemplateConfig
