+++ docs/examples/templating/engine_instance_minijinja.py
@@ -1,4 +1,4 @@
-from litestar.contrib.minijinja import MiniJinjaTemplateEngine
+from litestar.plugins.minijinja import MiniJinjaTemplateEngine
 from litestar.template.config import TemplateConfig
 
 template_config = TemplateConfig(engine=MiniJinjaTemplateEngine, directory="templates")
