+++ docs/examples/templating/returning_templates_minijinja.py
@@ -3,7 +3,7 @@ from __future__ import annotations
 from pathlib import Path
 
 from litestar import Litestar, get
-from litestar.contrib.minijinja import MiniJinjaTemplateEngine
+from litestar.plugins.minijinja import MiniJinjaTemplateEngine
 from litestar.response import Template
 from litestar.template.config import TemplateConfig
