+++ docs/examples/templating/template_engine_minijinja.py
@@ -1,7 +1,7 @@
 from pathlib import Path
 
 from litestar import Litestar
-from litestar.contrib.minijinja import MiniJinjaTemplateEngine
+from litestar.plugins.minijinja import MiniJinjaTemplateEngine
 from litestar.template.config import TemplateConfig
 
 app = Litestar(
