+++ docs/reference/plugins/minijinja.rst
@@ -1,0 +1,7 @@
+=========
+minijinja
+=========
+
+
+.. automodule:: litestar.plugins.minijinja
+    :members:
