+++ docs/release-notes/whats-new-3.rst
@@ -20,6 +20,8 @@ Imports
 +----------------------------------------------------+------------------------------------------------------------------------+
 | Put your changes here from v2                         | Put your changes here from v3                                       |
 +----------------------------------------------------+------------------------------------------------------------------------+
+| ``from litestar.contrib.minijinja import …``       | ``from litestar.plugins.minijinja import …``                           |
++----------------------------------------------------+------------------------------------------------------------------------+
 
 
 Removal of ``StaticFileConfig``
